fix(core): improved package size
