+++ apps/docs/app/(docs)/page.mdx
@@ -26,7 +26,6 @@ Now follow the usage instructions for your framework or builder:
 
 - [Next.js](/next)
 - [Vite](/vite)
-- [Create React App](/create-react-app)
 
 > **Important**: Before going to production, we recommend [enabling cache for the Bluesky API](#enabling-cache-for-the-bluesky-api) as server IPs might get rate limited by Bluesky.
 
@@ -36,7 +35,7 @@ The [`prefers-color-scheme`](https://developer.mozilla.org/en-US/docs/Web/CSS/@m
 
 ### Toggling theme manually
 
-The closest `data-theme` attribute on a parent element can determine the theme of the tweet. You can set it to `light` or `dark`, like so:
+The closest `data-theme` attribute on a parent element can determine the theme of the post. You can set it to `light` or `dark`, like so:
 
 ```tsx
 <div data-theme="dark">
@@ -66,6 +65,6 @@ For Global CSS the usage of `:global` is not necessary.
 
 ## Enabling cache for the Bluesky API
 
-Rendering posts requires making a call to Bluesky's syndication API. Getting rate limited by that API is very hard but it's possible if you're relying only on the endpoint we provide for SWR (`bsky-react-post.vercel.app/api/post`) as the IPs of the server are making many requests to the syndication API. This also applies to RSC where the API endpoint is not required but the server is still making the request from the same IP.
+Rendering posts requires making a call to Bluesky's syndication API. Getting rate limited by that API is very hard but it's possible if you're relying only on the endpoint we provide for SWR (`bsky-react-post.rhinobase.io/api/post`) as the IPs of the server are making many requests to the syndication API. This also applies to RSC where the API endpoint is not required but the server is still making the request from the same IP.
 
 To prevent this, you can use a db like Redis to cache the posts. If you're using Next.js then using [`unstable_cache`](/next#enabling-cache) works too.
