+++ packages/core/project.json
@@ -5,5 +5,11 @@
   "projectType": "library",
   "tags": [],
   "// targets": "to see all targets run: nx show project core --web",
-  "targets": {}
+  "targets": {
+    "nx-release-publish": {
+      "options": {
+        "packageRoot": "dist/{projectRoot}"
+      }
+    }
+  }
 }
