+++ packages/core/src/components/Post/utils.ts
@@ -1,0 +1,10 @@
+import type { AppBskyFeedPost } from "@atproto/api";
+import { hasProp, isObj } from "../../utils";
+
+export function isRecord(v: unknown): v is AppBskyFeedPost.Record {
+  return (
+    isObj(v) &&
+    hasProp(v, "$type") &&
+    (v.$type === "app.bsky.feed.post#main" || v.$type === "app.bsky.feed.post")
+  );
+}
